Split finding and car value objects
Extract the larger finding and car supporting types into focused domain modules, update package exports and external consumer imports to use vibesensor.domain, and refresh the domain hygiene/docs to match the new ownership boundaries.
