+++ apps/server/tests/adapters/pdf/_report_persistence_helpers.py
@@ -35,7 +35,7 @@ def build_findings(
 
 
 def findings_at_freq(findings: tuple | list, *freq_strs: str) -> list:
-    from vibesensor.domain.finding import Finding
+    from vibesensor.domain import Finding
 
     def _matches(finding: object) -> bool:
         if isinstance(finding, Finding):
