+++ apps/server/tests/adapters/pdf/test_report_analysis_helpers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from vibesensor.domain.finding import speed_band_sort_key, speed_bin_label
+from vibesensor.domain import speed_band_sort_key, speed_bin_label
 from vibesensor.shared.constants import KMH_TO_MPS
 from vibesensor.shared.json_utils import as_float_or_none as _as_float
 from vibesensor.shared.statistics_utils import (
