+++ apps/server/tests/hygiene/test_layer_boundaries.py
@@ -78,7 +78,7 @@ def _import_target_layer(module: str) -> str | None:
 _KNOWN_VIOLATIONS: frozenset[tuple[str, str]] = frozenset(
     {
         # domain → shared (root-level utilities)
-        ("domain/finding.py", "vibesensor.coerce"),
+        ("domain/finding_evidence.py", "vibesensor.coerce"),
         ("domain/location_hotspot.py", "vibesensor.coerce"),
         ("domain/order_match.py", "vibesensor.coerce"),
         ("domain/run_capture.py", "vibesensor.strength_bands"),
