+++ apps/server/tests/integration/test_multi_domain_regressions.py
@@ -14,8 +14,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from vibesensor.adapters.pdf.presentation import order_label_human
-from vibesensor.domain import Finding, SpeedSourceKind
-from vibesensor.domain.finding import speed_bin_label
+from vibesensor.domain import Finding, SpeedSourceKind, speed_bin_label
 from vibesensor.report_i18n import resolve_i18n as resolve_i18n_impl
 from vibesensor.report_i18n import tr
 from vibesensor.shared.json_utils import as_float_or_none as runlog_as_float_or_none
