+++ apps/server/tests/use_cases/diagnostics/test_analysis_pipeline_guard_regressions.py
@@ -20,7 +20,7 @@ from unittest.mock import patch
 
 import pytest
 
-from vibesensor.domain.finding import speed_bin_label
+from vibesensor.domain import speed_bin_label
 from vibesensor.use_cases.diagnostics.order_analysis import _compute_effective_match_rate
 from vibesensor.use_cases.run.logger import _run_loop
 from vibesensor.use_cases.updates.status import hash_tree
