+++ apps/server/tests/use_cases/run/conftest.py
@@ -14,7 +14,7 @@ from typing import Any
 
 import pytest
 
-from vibesensor.domain.car import CarSnapshot
+from vibesensor.domain import CarSnapshot
 from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
 from vibesensor.use_cases.run import RunRecorder, RunRecorderConfig
