+++ apps/server/vibesensor/shared/boundaries/vibration_origin.py
@@ -6,7 +6,7 @@ from collections.abc import Mapping
 from typing import TypedDict
 
 from vibesensor.coerce import coerce_float, coerce_int
-from vibesensor.domain.finding import Finding, VibrationSource
+from vibesensor.domain import Finding, VibrationSource
 from vibesensor.domain.location_hotspot import LocationHotspot
 from vibesensor.domain.vibration_origin import VibrationOrigin
 from vibesensor.shared.constants import PHASE_I18N_KEYS
