+++ apps/server/vibesensor/shared/order_bands.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from collections.abc import Mapping
 from math import isfinite, sqrt
 
-from vibesensor.domain.car import OrderReferenceSpec
+from vibesensor.domain import OrderReferenceSpec
 from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
 from vibesensor.shared.constants import (
     FREQUENCY_EPSILON_HZ,
