+++ apps/server/vibesensor/use_cases/diagnostics/order_heuristics.py
@@ -5,8 +5,7 @@ from __future__ import annotations
 from dataclasses import replace
 
 from vibesensor.domain import Finding as DomainFinding
-from vibesensor.domain import OrderMatchObservation
-from vibesensor.domain.finding import VibrationSource
+from vibesensor.domain import OrderMatchObservation, VibrationSource
 from vibesensor.shared.constants import ORDER_MIN_CONFIDENCE, ORDER_MIN_MATCH_POINTS
 from vibesensor.use_cases.diagnostics.math_utils import _mean
