+++ apps/server/vibesensor/use_cases/run/sample_builder.py
@@ -12,7 +12,7 @@ from dataclasses import asdict
 from typing import TYPE_CHECKING, NamedTuple, cast
 
 from vibesensor.coerce import coerce_float
-from vibesensor.domain.car import CarSnapshot
+from vibesensor.domain import CarSnapshot
 from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
 from vibesensor.domain.strength_metrics import StrengthMetrics
 from vibesensor.shared.constants import MPS_TO_KMH, NUMERIC_TYPES
